Refactor changes (check description)
All moments-forces moved except those that require rocket params or are time dependent. Other classes should ensure that they implement their own set_symbols() to ensure their symbols exist. Rocket params and env params are set to None. Efficiencies can be considered. Some notes in comments are left behind.
